Distributed PPO (#40)
Use the Hugging Face Accelerate library to support distributed training where separate processes are their own APPO training process that synchronizes at the forward-backward training steps using DistributedDataParallel

* Add accelerate library 

Using the higher-level library to accomplish distributed data parallel training

* DPPO uses accelerate library

DPPO is separate from APPO because it appears to be ~40% slower on local

* APPO uses optimized RolloutDataLoader

RolloutDataLoader doesn’t support multi-processes

* DPPO uses RolloutDataLoader

This appears to fix the biggest performance regression

* appo_benchmark.sh “supports” dppo

* PolicyActorPool.get_policy_for_actor_id takes cuda_index

When a workers actually takes a GPU, then it’s best to keep the inference actor on the same GPU because theoretically the GPU is free waiting for the inference actor.

* appo.yml -> dppo.yml

* DPPO LuxAI_S2-v0-j1024env32-a100-debug

* APPO: Need to move minibatch to correct device

* TEST: Keep the same dataloader in DPPO

Test to see if memory explosion still occurs

* Revert "TEST: Keep the same dataloader in DPPO"

This reverts commit 2c6ea36bfa43b7708238d1b011a7fd0c04acc363.

* DPPO: Clear dataloaders after done with rollouts

* Remove unused import

* LuxAI_S2-v0-j256env64-80m-ent1-lr01-mgn2-base1lc-ga2-tkl5-2lw-8rw-4a100

4 GPUs (2 gradient accumulators, 2 inference workers)

* appo_benchmark.sh defaults to both dppo and appo

* LuxAI_S2-v0-j256env64-4a100-debug

Skip prepare step to see if multi-GPU SGD is occurring

* Add debug statement for learner GPUs

* LuxAI_S2-v0-j256env64-4a100-debug derives from 4a100 job

Got the intial alias wrong

* Add Accelerator logging

* Fix typo

* Debug script for accelerate.Accelerator

* Set logging level to show logging.info

* debug_accelerate: Set CUDA_VISIBLE_DEVICES

* accelerate_mnist.py test script

* dppo -> dpppo

Mostly I’m thinking of this as Data Parallel PPO instead of Distributed PPO.

* TEST: APPO LunarLander-v2 using DataParalell

Seeing if this doesn’t simply crash

* eval_gpu_index -> evaluator_gpu_index

* inference_gpu_index -> inference_gpu_indexes

* env_id: LunarLander-v2

* Revert "env_id: LunarLander-v2"

This reverts commit c7bb3acabc61fdc8a52ed56908d88435e517064b.

* Revert "inference_gpu_index -> inference_gpu_indexes"

This reverts commit 066994eff1af7db3914bce2967635d8dd7ee3499.

* Revert "eval_gpu_index -> evaluator_gpu_index"

This reverts commit c17c2ee7c25447b73a05884804cb433a01d9a344.

* Revert "TEST: APPO LunarLander-v2 using DataParalell"

This reverts commit 4821d98ac0093737d6955d1be52581ea44c78af7.

* Try launching Accelerator instances from mp.spawn

* Specify address and port for process group

* Add torch.distributed.is_initialized check

Also move Accelerator import to top-level

* Print out visible cuda devices

* accelerator.device

* os.environ["LOCAL_RANK"] = str(rank)

* Try MNIST with debug_accelerate

* Add logs for dataset length, 5 epochs

* Send a serialized version of the model

* Try to serialize the model into and out of the worker

* set_obj -> get_obj

* evaluate model at beginning and end

* Fix reference

* Is this model actually being transferred?

* Print loss

* Don’t print accelerate state

* Use better model and log loss every epoch

* Don’t flatten the input

* Fix the SimpleCNN

* Use NLLLoss instead

* optimizer needs to be initialized from created model?

* Don’t optimizer.load_state_dict

* Actually use f-string

* Try to make eval faster

* Output to correct device

* Apparently models don’t have device by default

* Keep model on CPU for now

* Add more debugging on NaN

* Add softmax to model output

* F.log_softmax(x, dim=1)

* Revert "dppo -> dpppo"

This reverts commit 32621f30702ed2e48ec0a804c496bf05b6f27666.

* debug_accelerate -> mnist_accelerate

* dist_accelerate: What’s available to customize?

* Move Accelerator() up

* accelerator.use_distributed

* Print out gpu ids

* gpu.id

* Make DPPO compatible with `accelerate launch`

Each prcess is it’s own “APPO” job

* dppo uses accelerate launch

* Move minibatches to device

This didn’t have to be done when accelerate prepared the dataloader, but I’m using the dataset as is (since each process independenly generats its own rollouts.

* $launch not &launch

* Push the device transfer logic into RemoteInferencePolicy

The issue with self.policy = accelerator.unwrap_model(policy).to("cpu") is that it actually sets the accelerator training policy to cpu as well (at least in a non-distributed setting)

* In DPPO autocast_loss is handled by automate config

* dppo remove autocat_loss

LuxAI_S2-v0-j256env64-2a100-debug aliases the 2a100 job

* Use partial_state to synchronize on downloading model

* Remove unused import

* Remove unsupported policy functions

accelerate wraps the policy in a module, which has fewer capacilibites.

* Print out the actual device

* Don’t gather step metrics for each learner

These jobs are surprisingly separated, so this might be more indicative. More importantly, this is the more straightforward fix as I’m just seeing if this works at all.

* Update minibatches_per_step since datasets aren’t split

* Microrts-debug doesn’t use half-guided

half-guided and guided isn’t supported in the new rollout generation system

* Fix typo

* Don’t need MaskResettableEpisodeStatistics for vec_env

This is already tracked at the individual env level with RecordEpisodeStatistics

* LuxAI_S2-v0-j256env64-80m-ent1-lr01-mgn2-base1lc-ga4-tkl5c-4rw-h100

Like 2a100 but all on one GPU

* Disable CarRacing-v2 from appo_benchmark.sh

Need to fix saving sde

* Fix worker alias

+++ rl_algo_impls/runner/wandb_load.py
@@ -3,6 +3,8 @@ import os
 import shutil
 from typing import Tuple, TypeVar
 
+import accelerate
+
 import wandb
 from rl_algo_impls.runner.config import Config, Hyperparams, RunArgs
 
@@ -26,11 +28,14 @@ def load_player(
     config = Config(args, Hyperparams.from_dict_with_extra_fields(params), root_dir)
     model_path = config.model_dir_path(best=best, downloaded=True)
 
-    model_archive_name = config.model_dir_name(best=best, extension=".zip")
-    run.file(model_archive_name).download()
-    if os.path.isdir(model_path):
-        shutil.rmtree(model_path)
-    shutil.unpack_archive(model_archive_name, model_path)
-    os.remove(model_archive_name)
+    partial_state = accelerate.state.PartialState()
+    if partial_state.is_local_main_process:
+        model_archive_name = config.model_dir_name(best=best, extension=".zip")
+        run.file(model_archive_name).download()
+        if os.path.isdir(model_path):
+            shutil.rmtree(model_path)
+        shutil.unpack_archive(model_archive_name, model_path)
+        os.remove(model_archive_name)
+    partial_state.wait_for_everyone()
 
     return args, config, model_path
